+++ App/Screens/userPorfile.js
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import { StatusBar, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
+import { Image, StatusBar, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
 import { AppContext } from './../Context/AppContext';
 import { AuthContext } from './../Context/AuthContext';
 
@@ -18,6 +18,21 @@ const UserPorfile = ({ navigation }) => {
                 backgroundColor="#fff"
                 translucent={true}
             />
+             <View style={{
+                position: "absolute",
+                left: 10,
+                top: 20,
+            }}>
+                <TouchableOpacity onPress={() => navigation.toggleDrawer()}>
+                    <Image
+                        style={{
+                            height: 30,
+                            width: 30,
+                        }}
+                        source={require("../assets/icons/menu.png")}
+                    />
+                </TouchableOpacity>
+            </View>
             <View style={styles.details}>
                 <View>
                     <Text>Email: {userLogInInfo.data.email}</Text>