+++ App/ChangeLog.md
@@ -68,3 +68,8 @@
 
 - sub Catagory can be add.
 - Sub catagory update base builded.
+
+> Chnagelog: V-0.1.5:
+
+- Sub catagori can be updted.
+- Drawer toggle has been added.